requirements.txt
langchain>=0.1.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.1.0
mcp>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0 